Refactor array validation functions into cext_helpers.h.
